+++ OpenSim/Simulation/Model/Scholz2015GeometryPath.h
@@ -433,7 +433,27 @@ public:
     double getLengtheningSpeed(const SimTK::State& s) const override;
     double computeMomentArm(const SimTK::State& s,
             const Coordinate& coord) const override;
-    bool isVisualPath() const override { return true; }
+    bool isVisualPath() const override;
+
+    /**
+     * Find the list of paths to independent coordinates which fully determine
+     * the kinematic state of this path.
+     *
+     * `Scholz2015GeometryPath`'s concrete implementation of this method finds
+     * the joints lying between the frames associated with the path's origin and
+     * insertion points and returns the coordinate paths associated with these
+     * joints. Locked coordinates, prescribed coordinates, and coordinates
+     * dependent on other coordinates via a `CoordinateCouplerConstraint` are
+     * excluded from the list.
+     *
+     * @note This method uses several passes through the model's topology to
+     * form the list of coordinate paths, so avoid repeated calls in performance
+     * critical applications.
+     *
+     * @see SimulationUtilities::findJointsBetweenPhysicalFrames()
+     */
+    std::vector<ComponentPath>
+    findIndependentCoordinates(const SimTK::State&) const override;
     // @}
 
     //** @name `ForceProducer` interface */